Add category handling to authoring tool. Categories can be arbitrarily nested. However, titles have to be unique. If a category gets removed from the server side, the category setting is reset automatically. Not sure that is the best solution, but works for now.
Categories are meant to be managed on the server side, though it may be possible in
the future to add an "add category" function to the authoring tool side. The result
of this would be pushed to the server and therefore be available to everybody
using this instance.

Now categories are queried on startup - if they exist on the server and if login
succeeds. If no categories are available, the empty category is assigned automatically.